GEODE-7864: Remove redundant casts
Authored-by: Donal Evans <doevans@vmware.com>

+++ geode-apis-compatible-with-redis/src/main/java/org/apache/geode/redis/internal/netty/ByteToCommandDecoder.java
@@ -180,7 +180,7 @@ public class ByteToCommandDecoder extends ByteToMessageDecoder {
       }
       b = buffer.readByte();
       if (Character.isDigit(b)) {
-        number = number * 10 + (int) (b - '0');
+        number = number * 10 + (b - '0');
         readerIndex++;
       } else {
         buffer.readerIndex(readerIndex);
