+++ geode-core/src/main/java/org/apache/geode/cache/query/internal/index/RangeIndex.java
@@ -712,7 +712,7 @@ public class RangeIndex extends AbstractIndex {
               if (x < 0) {
                 x = 0;
               }
-              size = (int) x;
+              size = x;
             } else {
               // not attempting to differentiate between LT & LE
               size = this.valueToEntriesMap.containsKey(key) ? 1 : 0;
