+++ geode-core/src/main/java/org/apache/geode/internal/cache/tier/sockets/CacheServerHelper.java
@@ -153,7 +153,7 @@ public class CacheServerHelper {
                 "UTF-8 Exception malformed input",
                 new UTFDataFormatException("malformed input: partial character at end"));
           }
-          char2 = (int) bytearr[count - 1];
+          char2 = bytearr[count - 1];
           if ((char2 & 0xC0) != 0x80)
             throw new RuntimeException("malformed input around byte " + count);
           chararr[chararr_count++] = (char) (((c & 0x1F) << 6) | (char2 & 0x3F));
@@ -166,8 +166,8 @@ public class CacheServerHelper {
                 "UTF-8 Exception malformed input",
                 new UTFDataFormatException("malformed input: partial character at end"));
           }
-          char2 = (int) bytearr[count - 2];
-          char3 = (int) bytearr[count - 1];
+          char2 = bytearr[count - 2];
+          char3 = bytearr[count - 1];
           if (((char2 & 0xC0) != 0x80) || ((char3 & 0xC0) != 0x80)) {
             throw new RuntimeException(
                 "UTF-8 Exception malformed input",
